Improve htlc_maximum_msat in channel updates
We previously set the `htlc_maximum_msat` inside `channel_update` to the
channel's capacity, but that didn't make any sense: we will reject htlcs
that are above the local or remote `max_htlc_value_in_flight_msat`.

We now set this value to match the lowest `max_htlc_value_in_flight_msat`
of the channel, and properly type our local value to be a millisatoshi
amount instead of a more generic UInt64.

+++ eclair-core/src/main/scala/fr/acinq/eclair/channel/fsm/ChannelOpenDualFunded.scala
@@ -27,7 +27,7 @@ import fr.acinq.eclair.channel.fsm.Channel._
 import fr.acinq.eclair.channel.publish.TxPublisher.SetChannelId
 import fr.acinq.eclair.transactions.Scripts
 import fr.acinq.eclair.wire.protocol._
-import fr.acinq.eclair.{Features, RealShortChannelId}
+import fr.acinq.eclair.{Features, RealShortChannelId, UInt64}
 
 /**
  * Created by t-bast on 19/04/2022.
@@ -117,7 +117,7 @@ trait ChannelOpenDualFunded extends DualFundingHandlers with ErrorHandlers {
         commitmentFeerate = input.commitTxFeerate,
         fundingAmount = input.fundingAmount,
         dustLimit = input.localParams.dustLimit,
-        maxHtlcValueInFlightMsat = input.localParams.maxHtlcValueInFlightMsat,
+        maxHtlcValueInFlightMsat = UInt64(input.localParams.maxHtlcValueInFlightMsat.toLong),
         htlcMinimum = input.localParams.htlcMinimum,
         toSelfDelay = input.localParams.toSelfDelay,
         maxAcceptedHtlcs = input.localParams.maxAcceptedHtlcs,
@@ -153,7 +153,7 @@ trait ChannelOpenDualFunded extends DualFundingHandlers with ErrorHandlers {
             temporaryChannelId = open.temporaryChannelId,
             fundingAmount = d.init.fundingContribution_opt.getOrElse(0 sat),
             dustLimit = localParams.dustLimit,
-            maxHtlcValueInFlightMsat = localParams.maxHtlcValueInFlightMsat,
+            maxHtlcValueInFlightMsat = UInt64(localParams.maxHtlcValueInFlightMsat.toLong),
             htlcMinimum = localParams.htlcMinimum,
             minimumDepth = minimumDepth.getOrElse(0),
             toSelfDelay = localParams.toSelfDelay,
